implement auto-parent joining,  too
