+++ backend/cloud/teams.js
@@ -3,7 +3,7 @@
 // eslint-disable-next-line @typescript-eslint/no-var-requires
 const CONSTS = require("./consts.js");
 // eslint-disable-next-line @typescript-eslint/no-var-requires
-const { fetchMyTeams } = require("./utils.js");
+const { fetchMyTeams, joinTeam } = require("./utils.js");
 const { Team, Conversation } = CONSTS;
 
 const ROLES = [
@@ -47,28 +47,6 @@ async function myTeams(request) {
 Parse.Cloud.define("myTeams", myTeams, {
   requireUser: true
 });
-
-
-async function joinTeam(teamId, user) {
-  console.log(user.id, "wants to join", teamId);
-  const sessionToken = user.getSessionToken();
-  const team = await (new Parse.Query(Team)).get(teamId, {sessionToken});
-
-  if (!await team.isMember("members", user.id)) {
-    await team.applyForMembership(user);
-    // need to refresh after becoming a member to ensure we can see the teams we should autojoin
-    await team.fetch({sessionToken});
-  }
-
-  const autojoinTeams = team.get("autojoin") || [];
-  console.log(team, autojoinTeams);
-
-  if (autojoinTeams.length > 0) {
-    await Promise.all(autojoinTeams.map(async ptr =>
-      await joinTeam(ptr.id, user)
-    ))
-  }
-}
 
 Parse.Cloud.define("joinTeam", async (request) => {
